requirements.txt
Flask>=0.10.1
pyclips>=1.0.7.343